refactor: Combatant → Character 대규모 리팩토링 및 CombatStats 시스템 통합
주요 변경사항:
- Combatant → Character 전체 프로젝트 변경 (클래스명, 파일명, 참조)
- CombatantInventory → CharacterInventory 변경
- Character 클래스에 CombatStats 통합 (개별 필드 → 통합 구조체)
- CharacterData deprecated 필드 완전 제거

기능 추가:
- CharacterData에 초기 인벤토리 시스템 추가 (initialItems, initialEquipment)
- CharacterManager에서 모든 Character의 Inventory 자동 생성 및 초기화
- InventoryUI 자동 연결 시스템 (PlayerCharacter.Inventory)
- B 키로 인벤토리 토글 기능 구현
- InventoryTestManager TestMode 제어 추가

StatsCalculationManager 개선:
- RecalculateAndCommit() 수정: character.stats에 finalStats 커밋
- HP/Poise 비율 보존 로직 개선
- 네이밍 통일: GetEffective* → GetFinal*, Map* → Convert*

테스트 인프라:
- StatsTest.cs 작성 (9개 테스트 케이스 + 통합 테스트)
- 리팩토링_테스트_케이스.md 작성 (검증 절차서)

버그 수정:
- RecalculateAndCommit()가 character.stats에 커밋하지 않던 문제 수정
- InventoryUI가 PlayerCharacter.Inventory에 연결되지 않던 문제 수정
- Input Actions의 UI 맵이 자동 활성화되지 않던 문제 수정

영향받은 파일:
- Character.cs, CharacterData.cs, CharacterManager.cs
- StatsCalculationManager.cs, CombatStats.cs
- CharacterInventory.cs, InventoryUI.cs
- CombatManager.cs (1985줄), PlayerController.cs, EnemyController.cs
- BT 시스템 전체, AI 시스템 전체, UI 시스템 전체
- 총 80개 이상 파일 업데이트

컴파일 상태: 0 에러, 0 경고

+++ Assets/Script/UI/HPPanelController.cs
@@ -86,13 +86,13 @@ public class HPPanelController : MonoBehaviour
         if (CharacterManager.Instance?.PlayerCharacter != null)
         {
             currentPlayerHP = CharacterManager.Instance.PlayerCharacter.currentHP;
-            maxPlayerHP = CharacterManager.Instance.PlayerCharacter.MaxHP;
+            maxPlayerHP = (int)CharacterManager.Instance.PlayerCharacter.MaxHP;
         }
 
         if (CharacterManager.Instance?.EnemyCharacter != null)
         {
             currentEnemyHP = CharacterManager.Instance.EnemyCharacter.currentHP;
-            maxEnemyHP = CharacterManager.Instance.EnemyCharacter.MaxHP;
+            maxEnemyHP = (int)CharacterManager.Instance.EnemyCharacter.MaxHP;
         }
 
         if (debugMode)
